hist: Trim down memory requirements and sory by count
I misunderstood and was sorting by the line itself, which is much slower
on memory-constrained systems (thrashing). Sorting by count is what's
supposed to happen and it's far faster.

To make the hash table smaller, the hash is no longer stored, and the
count and length have been reduced to 32-bit integers.