cols: Add -p (padding) option
This simplifies a few things, and is conceptually how I should have
written it in the first place. Rather than track column padding as a
separate quantity, set the column width appropriately such that it's
always greater than the widest column and it works out naturally.

I wanted this feature for my personal thesaurus printout. Some of the
lines include phrases (words with spaces), which can be ambiguous when
columns are only separated by a single space.