Add uwords implementations
The actual tool isn't so useful, but some of the guts could be reused,
particularly the rune reader. The C implementation barely uses libc and
could be made freestanding with little difficulty.

The Go implementation is just for comparison. It makes for a much
simpler and shorter program since the built-ins and rich standard
library features do the heavy lifting, but the result has only a quarter
of the performance.

It may be worth using uint32_t for lengths/offsets — or perhaps even
uint16_t — trading off capability for better performance in the common
case. Real world inputs have few unique words, so these data structures
need not be so large.