pbmview: Refactor to support drag-and-drop
I was stumped about how to handle this correctly given the monitor
thread blocked on ReadDirectoryChangesW, and that the main thread must
never load files itself after startup. After sleeping on the problem, I
realized it could be solved by creating a new state and thread each time
the path changes, leaving state cleanup to the monitor thread.

This potentially leaks threads and states if the monitor threads are
never awoken from ReadDirectoryChangesW, but that's fine. It's not as
though someone is going to drag-and-drop *that* many files in a single
run for this to become an issue.

It also means the program must never return from main since at least one
monitor thread is still running.