qoi: Forbid repeated index operations
I had misunderstood the spec, and it specifically forbids repeat index
operations on the same index, not generally. I understand this better
after having written an encoder: The end-of-stream marker becomes a
reliable decoding error. The restriction also tightens the spec just
enough that there is always a canonical, optimal encoding.