pbmview: Correctly account for BMP padding
Rows in BI_RGB are 4-byte aligned, so images with widths not divisible
by 4 were displayed incorrectly. This padding is not mentioned in the
API documentation but is implied by using BMP.

As a result, give up on the idea of converting image formats "in place"
since this generally doesn't work correctly. Instead memory map the
input temporarily, then encode into an entirely new buffer.