cmdline.c: Fix surrogate pair following metachar
Metacharacter handling terminates with an "ungetc" which backs off the
read pointer so that the code point is re-read in the new state. When
the input was a surrogate pair, two code points were consumed, so it
must back off by two read positions to "ungetc" both surrogate halves.
Otherwise it retries with only half a surrogate pair, which besides
decoding incorrectly, causes certain valid UTF-16 inputs to decode as
WTF-8 instead of UTF-8 as promised.

Huge thanks to Fredrik Engkvist for catching and fixing this!
https://lists.sr.ht/~skeeto/public-inbox/%3C038e01d9176b%2427e04910%2477a0db30%24%40incrediblejunior.com%3E