canfigger.c: Do not modify the buffer in the parser
This leaves null terminating tokens as an option for the caller without
mandating it. The parser can now handle two new cases:

* Configs in a read-only buffer, such as a string literal or file map
* Parser can be restarted/rerun from an earlier state

The latter is demonstrated in a new example, -DEXAMPLE1, which was the
motivating situation. Though -DEXAMPLE2 shows how to accomplish the same
without it.

Casting away const is a feature, following the rule "accept general,
return specific." It is frequently regarded as a defect, but that is a
common mistake by language and interface designers, including even WG14
(and to a lesser extent WG21). Theoretically it lets the caller quietly
remove const from a pointer, but this is not a practical issue and the
trade-off is well worth it. Let's not forget that good engineering is
about trade-offs, not purity.

I wouldn't use const at all, except that I like two particular cases:

* As a storage qualifier for static data, improving binary image layout
* Communicating intent through the prototype ("promise not to mutate")