summarize.c: Compute destination before the loop
Due to a language design flaw, C compilers must allow "buf" to point at
"len". If that were to happen, each destination write would change the
destination of the next write, creating a loop-carried dependency that
prevents obvious optimizations. Of course this will not happen, but I
cannot communicate that constraint to the compiler. I wish I could
typedef a "Byte" that couldn't alias "len" but no C or C++ compiler
offers this feature.