summarize.c: Pass a count to FUTEX_WAKE
It's not documented, but it's easy to see in the Linux sources: The
minimum count for FUTEX_WAKE is 1 no matter what is passed, whether
that's zero or even negative. Since I only need to wake one waiter, it
doesn't actually matter whether or not I pass a count because any value
will work.