deflate.c: Use __stdcall and rework control flow
The "ums" functions use __stdcall, not __cdecl. Before this fix it was
crashing on 32-bit Windows when ums_inflate() was called more than a
couple of times, where it would gradually corrupt the stack.

Also rework the loop so that it handles zero-length inputs and such.
Namely, on Z_BUF_ERROR check avail_out to distringuish between "buffer
full" and "not enough input to make progress."