Add adafrien.c exercise solution
I'm particularly happy with this solver and generator, so I decided to
keep a more permanent copy. The original puzzle was probably not fully
thought out. Why limit costs to a billion? It's as if they wanted it to
fit in a 32-bit integer, except that totals may go as high as 47 bits,
and so a 64-bit integer (or double precision) is required. Better to
have limits costs up to 20000, then 32-bit integers are sufficient.