morsecode.c: Add a negative test (#5)
The decoder must immediately reject invalid input because (1) it allows
callers to detect which input was invalid and produce an accurate error
message, and (2) the state will eventually overflow for long inputs. The
change tests the particular case of 6 dots in a row, which should reject
on the 6th dot.