sortu64.c: Switch to base-128
The last digit was not handled correctly, resulting in a negative shift
that conceptually filled the lowest bits with zeros. This was hidden
from UBSan by insertion sort. That is, it never did radix sort on the
lowest digit in the tests because elements are too evenly distributed.

Rather than add a branch to check for negative shift and shift right
instead, handle the "spare" digit at the high end by shifting less the
first time. However, at base-32 this took a significant performance hit,
changing the optimum to base-128, which is even faster than before (in
GCC, Clang, and MSVC). This is surprising, because the top level sorts
only on the high bit, where it seems it wouldn't get much work done.