ptr.cpp: Don't use pointer arithmetic
My presumed case was objects allocated out of the same arena, but I took
a shortcut by testing on the stack. That made the pointer arithmetic
illegal, and so GCC optimized away most of the program. Computing with
integers makes this legal again so long as pointer-integer conversion is
straightforward.