rainbow.c: Use O(1) space computing chain merges
Instead of a hash table, discard the beginning of the longer chain, then
simply walk the chains side-by-side comparing hashes. This is O(1) space
and O(n+m) time. The previous version using a hash map was O(n) space,
and this was the primary memory and chain length constraint.

With this improvement, the program hardly uses memory at all and can run
practically indefinitely, churning out collisions at an ever increasing
rate. It's worth considering having a second shared hash table to track
discovered collsions in order to skip printing repeats.

Thanks to u/zookeeper_zeke for prompting the discussion that got me to
revisit this!