Add windows/du.c ("disk usage" for Windows)
Designed specifically around the Windows API in order to significantly
cut the expensive system calls that slow down du implementations. It
detects hard links using a probabilistic hash set of on-hand metadata,
cutting nearly all unnecessary CreateFile calls. In practice it's about
an order of magnitude faster than unix-native du ports.

Since this program doesn't spend much time waiting on system calls there
is little use in spinning up additional threads for concurrent waits. If
threads are added in the future, the queue could become a job queue, and
the hash-tries could be converted into concurrent hash-tries.

As is often the case with arena and hash-trie -oriented programs, there
is virtually no difference between -O0 and -O3. Bottlenecks are outside
the program. I suggest -O in the header just so that the generated code
is tidier.

I shamelessly went for GNU C on this one since that makes things much
simpler, but it still works perfectly well with MSVC/Clang. So no actual
dependency on GCC nor Mingw-w64.

It's interesting to contrast this with watc.c, which has nearly the same
core functionality around Find{First,Next}File. I've learned a lot since
writing that program, and I think it shows.