3sum.cpp: Adjust MSVC options, disable exceptions
It seems /kernel is the most reliable way to get what I want out of
MSVC, and even allows me to skip other workarounds. The changes also
make it work with the x86 compiler. (These notes will mostly serve as
reference documentation for future projects.)

The downside seems to be limited SIMD support (no AVX on x64, no SSE on
x86), though that doesn't matter for this program.