water-sort: Extract UI rendering and state from SDL
Emscripten looks like a disaster, so that's definitely out. Instead I've
worked out compiling directly to WASM using Clang and interfacing with
JavaScript, all 100% libc-free. With the UI platform-agnostic, I can run
it as WASM, passing inputs and receiving a DrawList for rendering to a
canvas. So it will all look and feel the same as the SDL version. I'll
probably use HTML buttons for reset, undo, etc. and other remaining UI
elements. The web UI will include with a sequence (10k?) of pre-selected
seeds ("levels") sorted by difficulty. No on-device puzzle generation,
which means no messing with workers and wasting battery life.

After seeing how the pieces fit together, it's as though I've been
practicing for freestanding WASM for years. Arena allocation and all
this do-it-yourself stuff is well-suited to WASM.