water-sort: Request animation frame first
The page randomly crashes on rare occasions when the tab is out of focus
in Firefox on one particular phone. There's no console, etc. on phones,
of course, so I can't see what's happening. I suspect it's a Firefox
bug. The Firefox on my Debian system fails to restore context, so I
already know Firefox WebGL is buggy at least on Linux.

Maybe the animation loop is dying? So let's try that.