birdgen: Add Win32 and WASM versions
My in-laws love playing this game when they visit. I put the text up on
a television screen with a large font so everyone can easily read them
from the couch. Then we all speculate about which is the real species,
announcing our guesses and scoring the results. When we all get it
wrong, or the name is interesting, we look up the bird to learn about
it.

So the web version has large text and auto-links Wikipedia on the reveal
so that I can more easily bring up photos and habitat range for all to
see on the screen.