birdgen: Minor fixups
In GNU C, restricted pointer arithmetic on null is allowed, just as in
standard C++, and it will be true for standard C in the near future. So
it's true of any compiler supporting #embed, and therefore I'm free to
do it here, as I already did with built-in memcpy.