findc.c: use nop instead of deferred linking
This was a chicken-or-egg situation where I added the IR so that I could
optimize, and it didn't occur to me to make generation simpler and rely
on the not-yet-written optimizer. Instead of these deferred links, just
jump to nop and rely on the optimizer to eliminate those nops later.